feat: renames default resolvers.ts file in generation template
Renames the generated resolvers file to graphbackResolvers.ts

Fixes #1174
