+++ templates/ts-apollo-fullstack/.graphqlrc.yml
@@ -27,7 +27,7 @@ extensions:
           outputPath: ./client/src/graphql
         graphback-resolvers:
           format: ts
-          outputPath: ./server/src/resolvers
+          outputPath: ./server/src/graphbackResolvers
   ## Knex DB Migration config
   dbmigrations:
     ## See knex.js for db specific config format
